finguard360/backend/requirements.txt
Flask==2.2.5
Flask_SQLAlchemy==3.0.3
PyJWT==2.8.0
psycopg2-binary==2.9.7
Werkzeug==2.2.3
python-dotenv==1.0.0
Faker==18.9.0
gunicorn==21.2.0
Flask-Limiter==2.8.0
Flask-Cors==4.0.0
reportlab==4.0.0
boto3==1.28.0
pytest==7.4.0
bcrypt==4.0.1
